config.h.win32.in: Fix and update entries
Make config.h.win32.in reflect on the entries that are in the
autotools/Meson builds more closely.

Also declare that we have fd_set, declare SIZEOF_[SIZE|SSIZE]_T and
SIZEOF_LONG_LONG for 32-bit and x64 builds properly.

Note that since the Visual Studio 2015/2017 CRT's vsnprintf() and
snprintf() are not compliant enough (though they are much closer than
before) for GLib's purporses, we still use the gnulib's implementations
of them, as with the pre-Visual Studio 2015 builds.