Fix bug: poll is hanging despite added new source.
This change makes the need_wakeup optimization a bit less eager to
prevent sending the wakeup signal. There were multiple bugs with the
intended optimization:

bug#1. A comment in g_main_context_prepare said "context->need_wakeup
is protected by LOCK_CONTEXT/UNLOCK_CONTEXT, so need not set it yet."
The problem is that in the middle of that function the lock is
released when the prepare functions are called. And that is during
iteration over the sources. So it could happen that a new source
arrives while iterating, right in that small window when the context
is not locked, but is not accounted for in the result of
g_main_context_prepare.

bug#2. In g_main_context_iterate, the following happens, with the context
being unlocked:

   step#1. first g_main_context_prepare
   step#2. then g_main_context_query a few times
   step#3. g_main_context_poll
   step#4. g_main_context_check

g_main_context_poll will only poll those file descriptors that the
previous two prepared and identified as pollable at the moment.
Even if we forget about bug#1, or even if turns out that luckily
the new sources arriving in that small window will always be iterated
over at the end, even then: if new sources arrive right after step#2,
but before step#3, then those new sources will not be considered for
polling in this round. If all the other sources are silent, this could
mean that it is hanging indefinitely.

The way I found this issue was that I wrote a game library, and quite
often it happened that the game froze until I started moving the mouse,
or generating any sort of gdk events.